payday/requirements.txt
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
Bootstrap-Flask==2.4.1
boto3==1.35.70
botocore==1.35.70
build==1.2.2.post1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.0
constantly==23.10.4
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
cryptography==44.0.2
daphne==4.1.2
dask==2024.12.1
dj-database-url==2.3.0
Django==5.1.3
django-ace==1.36.2
django-appconf==1.0.6
django-autocomplete-light==3.11.0
django-autoslug==1.9.9
django-celery-email==3.0.0
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-currentuser==0.8.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-health-check==3.18.3
django-htmx==1.21.0
django-json-widget==2.0.1
django-mathfilters==1.0.0
django-model-utils==5.0.0
django-money==3.5.3
django-phonenumber-field==8.0.0
django-qr-code==4.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.1
django-simple-history==3.7.0
django-slick-reporting==1.3.1
django-storages==1.14.4
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
dnspython==2.7.0
docker==7.1.0
docutils==0.21.2
email_validator==2.2.0
et_xmlfile==2.0.0
Faker==33.0.0
fastapi==0.115.6
Flask==3.1.0
flask-babel==4.0.0
Flask-Executor==1.0.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flower==2.0.1
fsspec==2024.12.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanize==4.12.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonfield==3.1.0
keyring==25.6.0
kombu==5.4.2
lazy_loader==0.4
locket==1.0.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
nh3==0.2.20
numpy==2.1.3
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
partd==1.4.2
phonenumbers==8.13.50
pillow==11.0.0
pipdeptree==2.16.2
pkginfo==1.12.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psutil==6.1.1
psycopg2-binary==2.9.10
py-moneyed==3.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.12.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
readme_renderer==44.0
redis==5.2.0
reflex-chakra==0.6.2
reflex-hosting-cli==0.1.32
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
s3transfer==0.10.4
segno==1.6.1
sentry-sdk==2.19.0
service-identity==24.2.0
setuptools==75.6.0
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
sqlparse==0.5.2
starlette==0.41.3
starlette-admin==0.14.1
swapper==1.4.0
tabulate==0.9.0
tenacity==9.1.2
text-unidecode==1.3
tomlkit==0.13.2
toolz==1.0.0
tornado==6.4.2
twine==6.0.1
Twisted==24.11.0
txaio==23.1.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
validators==0.34.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.8.2
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
XlsxWriter==3.2.0
zope.event==5.0
zope.interface==7.1.1
channels-redis
django-redis
git+https://github.com/tabarochristian/django-notifications.git